+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Update Bitmovin's native iOS SDK version to [`3.116.0`](https://developer.bitmovin.com/playback/docs/release-notes-ios#31160)
+
 ## [1.21.0] - 2026-06-26
 
 ### Added